Several changes to SplinePath
- Move SplinePath.cs to SharpNeedle/SonicTeam
- Rename Aabb to Bounds
- Make DoubleKnot, UnknownStruct, UserData subclasses of PathObject
- Convert DoubleKnot to a struct instead of an IBinarySerializable class
